+++ 03_next_trip/src/components/ImageUpload.tsx
@@ -3,9 +3,11 @@ import AddImage from './../assets/add.svg';
 
 export default function ImageUpload() {
   return (
-    <div className='flex h-40 w-40 flex-col items-center justify-center gap-2 overflow-hidden rounded-lg border-none bg-[#f2f2f2] px-2 text-base font-normal leading-normal'>
+    <div className='h-25 w-25 flex flex-col items-center justify-center gap-2 overflow-hidden rounded-lg border-none bg-[#f2f2f2] px-2 text-base font-normal leading-normal tablet:h-40 tablet:w-40'>
       <Image src={AddImage} alt='addImage' />
-      <p>클릭해서 사진 업로드</p>
+      <p>
+        <span className='hidden'>클릭해서 </span>사진 업로드
+      </p>
     </div>
   );
 }
